requirements-render.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
requests==2.32.3